Move InternelaTests to AutoMockingTest namespace
