+++ test/Tethos.FakeItEasy.Tests/AutoMockingTest/InternalTests.cs
@@ -1,14 +1,15 @@
-﻿namespace Tethos.Moq.Tests
+﻿namespace Tethos.FakeItEasy.Tests.AutoMockingTest
 {
     using AutoFixture.Xunit2;
-    using Castle.DynamicProxy.Generators;
+    using Castle.MicroKernel;
     using FluentAssertions;
-    using global::Moq;
+    using global::FakeItEasy;
+    using global::FakeItEasy.Core;
     using Tethos.Extensions;
     using Tethos.Tests.Common;
     using Xunit;
 
-    public class InternalTests : AutoMockingTest
+    public class InternalTests : FakeItEasy.AutoMockingTest
     {
         [Theory]
         [AutoData]
@@ -17,10 +18,8 @@
         {
             // Arrange
             var sut = this.Container.Resolve<InternalSystemUnderTest>();
-
-            this.Container.Resolve<Mock<IMockable>>()
-                .Setup(mock => mock.Do())
-                .Returns(expected);
+            var mock = this.Container.Resolve<IMockable>();
+            A.CallTo(() => mock.Do()).Returns(expected);
 
             // Act
             var actual = sut.Do();
@@ -36,10 +35,8 @@
         {
             // Arrange
             var sut = this.Container.Resolve<SystemUnderTestWithInternal>();
-
-            this.Container.Resolve<Mock<IInternalMockable>>()
-                .Setup(mock => mock.Do())
-                .Returns(expected);
+            var mock = this.Container.Resolve<IInternalMockable>();
+            A.CallTo(() => mock.Do()).Returns(expected);
 
             // Act
             var actual = sut.Do();
@@ -56,7 +53,7 @@
             var sut = () => this.Container.Resolve<InternalDependency.Tests.SystemUnderTest>();
 
             // Act & Assert
-            sut.Should().Throw<GeneratorException>();
+            sut.Should().Throw<FakeCreationException>();
         }
 
         [Fact]
@@ -64,21 +61,21 @@
         public void ResolveFrom_LooseInternalDependency_ShouldThrowGeneratorException()
         {
             // Arrange
-            var sut = () => this.Container.ResolveFrom<InternalDependency.Tests.SystemUnderTest, Mock<InternalDependency.Tests.IMockable>>();
+            var sut = () => this.Container.ResolveFrom<InternalDependency.Tests.SystemUnderTest, InternalDependency.Tests.IMockable>();
 
             // Act & Assert
-            sut.Should().Throw<GeneratorException>();
+            sut.Should().Throw<FakeCreationException>();
         }
 
         [Fact]
         [Trait("Category", "Integration")]
-        public void Exercise_LooseInternalDependency_ShouldMatch()
+        public void Exercise_LooseInternalDependency_ShouldThrowComponentNotFoundException()
         {
-            // Act
-            var sut = this.Container.Resolve<Mock<InternalDependency.Tests.IMockable>>();
+            // Arrange
+            var sut = () => this.Container.Resolve<InternalDependency.Tests.IMockable>();
 
-            // Assert
-            sut.Should().NotBeNull();
+            // Act & Assert
+            sut.Should().Throw<ComponentNotFoundException>();
         }
     }
 }
